+++ engine/server/sv_game.cpp
@@ -348,7 +348,7 @@ int GameImport_FS_GetFileList( const char * path, const char * extension, char *
 
 }
 int GameImport_FS_Seek( int f, int64_t offset, int origin ) {
-	return FS_VM_SeekFile( f, offset, (fsOrigin_t)origin, H_QAGAME );
+	return FS_VM_SeekFile( f, (long)offset, (fsOrigin_t)origin, H_QAGAME );
 
 }
 void GameImport_SendConsoleCommand( int exec_when, const char * text ) {
