+++ game/game/g_client.cpp
@@ -222,7 +222,7 @@ gentity_t *SelectRandomFurthestSpawnPoint ( vec3_t avoidPoint, vec3_t origin, ve
 	}
 
 	// select a random spot from the spawn points furthest away
-	rnd = random() * (numSpots / 2);
+	rnd = (int)( random() * (numSpots / 2) );
 
 	VectorCopy (list_spot[rnd]->s.origin, origin);
 	origin[2] += 9;
